refactor confetti and decisions to client components and remove empty tags
